opensky_etl/requirements.txt
requests==2.31.0
pandas==2.1.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
psycopg2-binary==2.9.7
jinja2==3.1.2
pytest==7.4.0
pytest-cov==4.1.0